add vertical carousel story
