+++ apps/storybook/src/config/example.ts
@@ -1,0 +1,56 @@
+export const membersConfig = [
+  {
+    src: 'https://images.unsplash.com/photo-1544005313-94ddf0286df2?&w=150&h=150&dpr=2&q=70&crop=faces&fit=crop',
+    name: 'Keonhee',
+    email: 'keonhee.dev@example.com',
+    fallback: 'KH',
+    activity: {
+      action: 'Updated client details for issue #7',
+      activeAt: '10:12 am',
+    },
+  },
+  {
+    src: 'https://images.unsplash.com/photo-1522075469751-3a6694fb2f61?&w=150&h=150&dpr=2&q=70&crop=faces&fit=crop',
+    name: 'Jaesang',
+    email: 'borderjae@example.com',
+    fallback: 'JC',
+    activity: {
+      action: 'pushed a PR to dev branch',
+      activeAt: '11:34 am',
+    },
+  },
+  {
+    src: 'https://images.unsplash.com/photo-1526510747491-58f928ec870f?&w=150&h=150&dpr=2&q=70&crop=focalpoint&fp-x=0.48&fp-y=0.48&fp-z=1.3&fit=crop',
+    name: 'YoungJin',
+    email: 'newhar@example.com',
+    fallback: 'YK',
+    activity: {
+      action: 'Responded to your comment #7514',
+      activeAt: '13:04 pm',
+    },
+  },
+  {
+    src: 'https://images.unsplash.com/photo-1541823709867-1b206113eafd?&w=150&h=150&dpr=2&q=70&crop=focalpoint&fp-x=0.5&fp-y=0.3&fp-z=1.5&fit=crop',
+    name: 'Taekyung',
+    email: 'taekyung@example.com',
+    fallback: 'TS',
+    activity: { action: 'Created a new issue', activeAt: '15:14 pm' },
+  },
+  {
+    src: 'https://images.unsplash.com/photo-1532073150508-0c1df022bdd1?&w=150&h=150&dpr=2&q=70&crop=focalpoint&fp-x=0.48&fp-y=0.35&fp-z=2&fit=crop',
+    name: 'Youngyun',
+    email: 'yougyun@example.com',
+    fallback: 'YA',
+    activity: {
+      action: 'Approved pull request #361',
+      activeAt: '20:50 pm',
+    },
+  },
+  {
+    src: 'https://plus.unsplash.com/premium_photo-1691622500373-da5adfe5b2ff?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=687&q=80',
+    name: 'SeHee',
+    email: 'shinsehantan@example.com',
+    fallback: 'SS',
+    activity: { action: 'Joined the team', activeAt: '21:00 pm' },
+  },
+];
