+++ apps/storybook/src/config/index.ts
@@ -1,0 +1,1 @@
+export * from './example';
