+++ package.json
@@ -5,6 +5,7 @@
     "dev": "turbo run dev",
     "lint": "turbo run lint",
     "test": "turbo run test",
+    "clean": "rm -r apps/web/.next",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "prepare": "husky install",
     "publish-packages": "turbo run lint && changeset version && changeset publish",
